Create NCCLibrary.js
Create a JavaScript library to handle the loading of 26 traits from 26 inscriptions. This library should have functions to retrieve and organize the traits for generating random PFP art.

Write a function within the JavaScript library that generates random art based on the loaded traits. This function can use a combination of the traits to create unique and varied PFP art.